+++ bsl/externals/__init__.py
@@ -5,13 +5,7 @@ import importlib
 
 from ..utils._logs import logger
 
-try:
-    pylsl = importlib.import_module(name="pylsl")
-    logger.debug("Installed 'pylsl' is functional.")
-except (ModuleNotFoundError, ImportError, RuntimeError):
-    pylsl = importlib.import_module(name=".pylsl", package=__name__)
-    logger.debug("Module 'pylsl' has been imported from externals.")
-
+pylsl = importlib.import_module(name=".pylsl", package=__name__)
 
 try:
     psychopy = importlib.import_module(name="psychopy")
