+++ bsl/externals/pylsl/__init__.py
@@ -1,28 +1,10 @@
 # (not using import * for Python 2.5 support)
 from .pylsl import (
-    InternalError,
-    InvalidArgumentError,
-    LostError,
     StreamInfo,
     StreamInlet,
     StreamOutlet,
-    TimeoutError,
-    XMLElement,
-    cf_double64,
-    cf_float32,
-    cf_int8,
-    cf_int16,
-    cf_int32,
-    cf_int64,
-    cf_string,
     library_version,
     local_clock,
-    proc_ALL,
-    proc_clocksync,
-    proc_dejitter,
-    proc_monotonize,
-    proc_none,
-    proc_threadsafe,
     protocol_version,
     resolve_streams,
 )
